+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* README updates.
+* Add focusedBorderColor param.
+* Add focusedErrorBorderColor param.
+
 ## 1.0.0
 
 * Initial release.